CRD generation improvements re documentation & required fields
Updated Go code so that CRD Go struct fields' comments will become
documentation for their corresponding CRD fields. In other words, for a
CRD property, the comments of the corresponding Go struct field will be
the CRD property's documentation.

In addition, removed the "omitempty" JSON annotation from some of the
Go struct fields, as that makes them optional (for some fields that was
not the correct setting).