Infer next semantic version for builds
This change detects the next semantic version of the build, based on
past changes in the current branch from the last tag, and injects that
version into the tests and the Docker image, as well as adding that
version as a Docker tag to that image.

As part of this change, the workflow now tests and builds the docker
image in parallel, rather than in sequence (this makes Docker images of
builds that do not pass the tests available for usage).